move builtinfunction type back
